update docs, update/clean examples
