+++ examples/README.md
@@ -1,0 +1,5 @@
+# Examples
+
+This directory contains a number of simple examples for using the `node-pxgrid` module.
+
+For the sake of consistency and brevity, the setup of the module is handled once in pxgrid-setup.js and imported into each of the examples. This file serves as a good example of how to setup the module.
