+++ examples/playground.js
@@ -1,20 +1,7 @@
 // This file is just a testing playground for all
 // the pxgrid - control / client lib functions!
 // aka a total free for all!
-const certs = require('./certs.js');
-
-const Pxgrid = require('../');
-
-const pxgridControlOptions = {
-  host: 'dnaise.ironbowlab.com',
-  client: 'my-node-app',
-  clientCert: certs.clientCert,
-  clientKey: certs.clientKey,
-  caBundle: certs.caBundle,
-  clientKeyPassword: 'Pxgrid123'
-};
-const pxgrid = new Pxgrid.Control(pxgridControlOptions);
-const pxclient = new Pxgrid.Client(pxgrid);
+const { pxgrid, pxclient } = require('./pxgrid-setup');
 
 const ancCallback = function(message) {
   //console.log(message)
@@ -143,4 +130,4 @@ function main() {
   */
 }
 
-pxgrid.activate().then(() => main());
+pxclient.connect().then(() => main());
