+++ examples/publisher_1100000001.js
@@ -1,5 +1,6 @@
 const { pxgrid, pxclient } = require('./pxgrid-setup');
 
+const deviceType = process.argv[2] || 'Test Device Type';
 const body = {
   assetId: '260',
   assetName: 'Abjergaryn - 47',
@@ -10,20 +11,14 @@ const body = {
   assetProductId: 'IE2000',
   assetSerialNumber: '1212121213243',
   assetMacAddress: '11:00:00:00:00:01',
-  assetDeviceType: "Ryan Wolfe's Special Device",
+  assetDeviceType: deviceType,
   assetIpAddress: '1.2.3.4',
-  assetCustomAttributes: [
-    {
-      value: 'SuperDevice',
-      key: 'WolfeAttr'
-    }
-  ],
   assetConnectedLinks: []
 };
 
-pxclient.connect().then(session =>
+pxclient.connect().then(session => {
   pxclient.createEndpointAssetPublisher().then(() => {
     console.log('PUBLISHER ESTABLISHED');
     pxclient.publishEndpointAssetUpdate(session, body);
-  })
-);
+  });
+});
