+++ index.js
@@ -1,6 +1,6 @@
-const Pxgrid = {}
+const Pxgrid = {};
 
-Pxgrid.Control = require('./lib/control')
-Pxgrid.Client = require('./lib/client')
+Pxgrid.Control = require('./lib/control');
+Pxgrid.Client = require('./lib/client');
 
-module.exports = Pxgrid
+module.exports = Pxgrid;
