+++ lib/client.js
@@ -36,8 +36,8 @@ const ENDPOINT_ASSET_SERVICE = 'com.cisco.endpoint.asset';
  * const pxgrid = new Pxgrid.Control(pxgridControlOptions);
  * const pxclient = new Pxgrid.Client(pxgridControlObject);
  *
- * pxgrid.activate()
- *   .then(() => {
+ * pxclient.connect()
+ *   .then(session => {
  *     pxclient.getProfiles()
  *        .then(profiles => console.log(profiles));
  *
@@ -66,8 +66,8 @@ const ENDPOINT_ASSET_SERVICE = 'com.cisco.endpoint.asset';
  * const pxcontrol = new Pxgrid.Control(pxgridControlOptions);
  * const pxclient = new Pxgrid.Client(pxgridControlObject);
  *
- * pxcontrol.activate()
- *   .then(() => {
+ * pxclient.connect()
+ *   .then(session => {
  *     pxclient.getProfiles()
  *        .then(profiles => console.log(profiles));
  *
@@ -216,23 +216,21 @@ class Client {
    *
    * This may change in the future for better usage (e.g. wait period executed within connectToBroker()).
    *
-   * @param {boolean} [debug=false] - Enable debugging of publish/subscribe events going through the broker.
+   * @param {boolean} [debug=false] - Enables debugging of publish/subscribe events going through the broker.
    * @return {Promise} A STOMP client object.
    * @memberof Client
    *
+   * @see Client#connect
+   *
    * @example
    * const ancCallback = function(message) {
-   *   const body = JSON.parse(message.body);
-   *   console.log('Endpoint ' + body.macAddress + ' has had an ' + body.status + ' ANC event');
-   * }
+   *   const body = message.body;
+   *   console.log(`${Date.now()}: Endpoint ${body.macAddress} has had an ${body.status} ANC event`);
+   * };
    *
-   * function main() {
-   *   pxclient.connectToBroker()
-   *     .then(session => pxclient.subscribeToAncPolicies(session, ancCallback));
-   * }
-   *
-   * pxgrid.activate()
-   *   .then(() => main());
+   * pxclient
+   *   .connect()
+   *   .then(session => pxclient.subscribeToAncPolicies(session, ancCallback));
    */
   connectToBroker(debug = false) {
     return this._getServiceWithSecret(PUBSUB_SERVICE)
@@ -280,6 +278,21 @@ class Client {
       .catch(error => {
         throw new Error(error);
       });
+  }
+
+  // This is just for a cleaner call to connect the instance. Nothing special here.
+  /**
+   * @description
+   * Alternative to Client.connectToBroker().
+   * Primarily added to give a jargon-free method to connect.
+   *
+   * @param {boolean} [debug=false] Enables debugging.
+   * @return {Promise} A STOMP client object.
+   * @memberof Client
+   * @see Client#connectToBroker
+   */
+  connect(debug = false) {
+    return this.connectToBroker(debug);
   }
 
   // Session directory functions
@@ -962,7 +975,7 @@ class Client {
    * @param {string} assetBody.assetConnectedLinks - Arbitrary value.
    * @param {Object} assetBody.assetConnectedAttributes - Any other custom attributes that have been created in Cisco ISE.
    * @param {string} assetBody.assetCustomAttributes.attrName - AttrName should match name of custom attribute in ISE. Value is arbitrary.
-   * @param {boolean} [debug=false] - If true, will enable debug messages being logged to console.
+   * @param {boolean} [debug=false] - Enables debug messages being logged to console.
    * @memberof Client
    */
   publishEndpointAssetUpdate(stompClient, assetBody, debug = false) {
@@ -1040,19 +1053,15 @@ class Client {
    * @param {Object.<stompClient>} stompClient - The active broker session to use for subscription.
    * @param {string} service The custom service name that is providing the topic.
    * @param {string} topic The custom topic.
-   * @param {boolean} [debug=false]
+   * @param {boolean} [debug=false] Enables debugging.
    * @return {Object.publisher} A publisher object.
    * @memberof Client
    *
    * @example
    * pxclient
-   *  .connectToBroker()
-   *  .then(session =>
-   *    pxclient.createCustomPublisher(session, 'my.service.name', 'myTopic')
-   *  )
-   *  .then(publisher => {
-   *    publisher.publish({ someData: 'Some sort of data.' });
-   *  });
+   *  .connect()
+   *  .then(session => pxclient.createCustomPublisher(session, 'my.service.name', 'myTopic'))
+   *  .then(publisher => publisher.publish({ someData: 'Some sort of data.' }));
    */
   createCustomPublisher(stompClient, service, topic, debug = false) {
     return this._getTopicPublisher(service, topic).then(() => {
@@ -1089,22 +1098,13 @@ class Client {
    * @param {string} service The custom service name that is providing the topic.
    * @param {string} topic The custom topic.
    * @param {function(messageObject)} messageCallback - A callback function that handles the message coming in on a topic subscription.
-   * @param {boolean} [debug=false]
+   * @param {boolean} [debug=false] Enables debugging.
    * @memberof Client
    *
    * @example
-   * pxgrid.activate().then(() => {
-   *  pxclient
-   *   .connectToBroker()
-   *   .then(session =>
-   *     pxclient.subscribeToCustom(
-   *       session,
-   *       'blah.blah.blah',
-   *       'customTopic',
-   *       genericCallback
-   *     )
-   *   );
-   * });
+   * pxclient
+   *   .connect()
+   *   .then(session => pxclient.subscribeToCustom(session, 'blah.blah.blah', 'customTopic', genericCallback));
    */
   subscribeToCustom(
     stompClient,
