+++ lib/control.js
@@ -251,7 +251,7 @@ class Control {
   /**
    * @description Reregister your node for a service. Services must reregister within the reregisterTimeMillis interval provided when initially registering.
    * @memberof Control
-   * @see serviceRegister
+   * @see Control#serviceRegister
    * @param  {string} serviceId - The ID of the service to reregister.
    * @return {Promise} Empty response from controller, if successful.
    */
@@ -262,7 +262,7 @@ class Control {
   /**
    * @description Automatically reregister a service at a given interval. Reregistration will occur 5 seconds prior to provided interval to prevent inadvertent timeouts.
    * @memberof Control
-   * @see {@link serviceReregister}
+   * @see Control#serviceReregister
    * @param {string} serviceId - The ID of the service to reregister.
    * @param {number} interval - Interval to reregister (milliseconds).
    */
